[Bug] Fixed rendering context queue growing indefinitely
This is not really a bug, however, when Blend2D is used to render millions
of paths with MT rendering enabled, it would also allocate memory for all
edges of these paths, etc... This changeset makes sure that an internal
limit is applied if not overridden by the user. The internal limit is
currently 10240 commands (BL_RASTER_CONTEXT_DEFAULT_COMMAND_QUEUE_LIMIT).

Also, another minor change is to try to align worker arena allocators
after job processing to make sure that other threads would not share a
cache line in case that a worker allocated memory for a job and then uses
memory from the same cache line for an active edge.

Lastly, this changeset makes the rendering context allocate from a zeroed
memory pool when an image is attached to it, instead of postponing the
allocation. The reason is that in 99.9% cases users would want to render
a path, and having the buffer preallocated is a big win as it removes one
point of failure in a worker thread.