[Bug] Fixed BLPath::addStrokedPath() bug (invalid access)
This fixes a stroking bug that was only triggered by using a custom
path sink that appends the stroked path into the "a" path of the
sink. This functionality was only used b BLPath::addStrokedPath()
at the moment, so BLContext::strokeXXX() functionality is unaffected.

In addition to this issue, all stroked paths are now properly closed,
which only affects additional path processing after its stroked.