[Bug] Fixed a bug in rasterengine RGBA32, RGBA64, and RGBA color dispatching
This bug was not really triggered before as it requires a floating point
RGBA color to be assigned to the rendering context, which is not what the
majority users do.

In addition, this changeset also makes portable pipeline always compiled in
to prevent situations in which a JIT runtime cannot be created and Blend2D
would not be able to render anything without it.