Implemented JIT pipeline generator abstractions (GP instructions)
This changeset makes it closer to have a universal JIT compiler
in the pipeline generator where most of the logic can be portable.
To achieve this, PipeCompiler must abstract commons of each target
architecture into interface that is then used by pipeline parts
without depending on target architecture specifics.

This commit abstracts general purpose instructions that are used
to hold and operate on integer registers and pointers. Vector
operations are not yet abstracted.