[Bug] Fixed a portable TT glyph decoder
The problem was with flags repeat feature not applying it correctly
in case that the repeat contained an off point curve in cases in
which the previous point was not off point as well. Sometimes quads
were handled as lines in this case.

Ironically, this bug was only in the portable decoder -  SSE4.2 and
AVX2 optimized decoders applied these flags correctly.