[Bug] Fixed out of worker stack space in debug builds
This was never a problem in release builds, because compilers are
smart about stack use in this case. However, this is a problem in
debug builds, because when C++ compilers inline functions each value
gets its own stack slot, which is never reused in the same function.
Thus, when forcing inlining even in debug mode the stack of such
functions explode.

  * Smarter use of BL_INLINE_NODEBUG and BL_INLINE_IF_NOT_DEBUG
  * Added fuzz tests that fill text to actually have this covered