[Bug] Increase worker stack size in debug builds
This change was made to make ASAN builds working with multi-threaded
renderer. The problem is that 128kB worker stack size is too small
for ASAN builds, which would exhaust this stack in heavily inlined
SIMD code when compiled without optimizations (debug).

Many functions are marked as BL_INLINE_IF_NOT_DEBUG for this reason,
however, there are some SIMD implementations that use a lot of
abstractions (notebly SIMD optimized truetype decoder) and
unfortunately compilers use large stack space when these functions
are compiled in debug mode.

This is not a problem of release builds, so they keep using 128kB.