[bug] Fixed a border-case of a thread pool destruction
When Blend2D library shuts down (static destructors are called)
the thread pool instance gets destroyed too. There is an internal
time-out that lets the destructor wait for all threads to get
successfully released. However, when this timeout is reached and
a thread is still running the thread pool gets destroyed anyway.

In this case the problem was that a thread that is shutting down
was still trying to notify the thread pool about that, but the
thread pool was already destroyed.

This change fixes this case. There is now an internal reference
count that is used by each thread and when it gets zero the thread
pool gets destroyed and thread-pool memory released.