[bug] Fixed a too strict check when loading OTF (CFF) fonts
Some fonts set private dictionary offset to the end of the font
data even when it should be zero when it's not provided. In this
case Blend2D refused to process such fonts as it thought that the
data was corrupted / malformed.

This commit changes the logic so private offset at the end of the
file is allowed, but its size must be zero.