Optimized multi-threaded processing of small-size commands
Multi-threaded rendering context's command processor was processing
each command until the command's Y1 position (last scanline) was
no reachable anymore. This was good for commands that described
rendering of something close to the top, however, it bottlenecked
all commands that were in the middle or close to the bottom.

A new approach was developed to improve the performance of the
command processor. To quickly skip commands that are not to be
processed yet, each command queue now contains an array of
quantized coordinates that describe the start of the command
(top Y coordinate), but to fit into 8 bits it's quantized so
the extra buffer is not too big.

The quantization shift is determined when the rendering context
is created, and it describes the number of bits to discard to
always get a valid quantized coordinate between 0 to 254. 255
is reserved for "not initialized" or "not to render" in some
situations.

The command processor simply uses SIMD to calculate which commands
won't be processed based on the band coordinates in addition to the
existing pending mask. The bit-scanning is still used to quickly
skip commands.