Use kern table if there is no kerning in GPOS
There are fonts that provide both kern and GPOS tables, however,
the GPOS table provided doesn't have kern feature, thus the font
expects that the software would use both, kern and GPOS tables to
position glyphs.

This adds a better logic to decide whether to use kern table or not,
depending on the presence of GPOS table and kern feature in that
table.